@@ -8,7 +8,7 @@ buildscript {
         maven { url "http://jitpack.io" }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.4.0'
+        classpath 'com.android.tools.build:gradle:3.5.2'
     }
 }
 
@@ -25,10 +25,15 @@ task clean(type: Delete) {
 }
 ext {
     minSdkVersion = 21
-    targetSdkVersion = 28
-    compileSdkVersion = 28
-    buildToolsVersion = "28.0.3"
-    supportLibraryVersion = '28.0.0'
+    targetSdkVersion = 29
+    compileSdkVersion = 29
+    buildToolsVersion = "29.0.2"
+    appcompat = '1.1.0'
+    cardview = '1.0.0'
+    recyclerview = '1.1.0-rc01'
+
+
+    cardview
 //    glideVersion = '4.0.0'
 }
 
