@@ -1,8 +1,9 @@
 package com.snowy.autoscrollviewpager;
 
-import android.support.v4.view.PagerAdapter;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.viewpager.widget.PagerAdapter;
 
 public class AutoScrollPagerAdapter extends PagerAdapter {
 
