@@ -1,7 +1,6 @@
 package com.snowy.ttword.adapter;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
 import android.text.SpannableString;
 import android.text.Spanned;
 import android.text.TextPaint;
@@ -15,6 +14,8 @@ import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
 
 import com.snowy.common.utils.ScreenUtils;
 import com.snowy.ttword.R;
