@@ -1,11 +1,12 @@
 package com.snowy.ttword.adapter;
 
 import android.graphics.drawable.AnimationDrawable;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.snowy.ttword.R;
 
