@@ -1,13 +1,14 @@
 package com.snowy.ttword.fragment;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 
 import com.snowy.common.activity.BaseActivity;
 import com.snowy.common.dialog.BottomSlideDialog;
