@@ -10,8 +10,9 @@ import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.Color;
 import android.os.Build;
-import android.support.v4.app.NotificationCompat;
 
+
+import androidx.core.app.NotificationCompat;
 
 import com.snowy.ttword.R;
 
